refactor: move `ElementPerformFullscreenEnter` struct to `fullscreen/element.rs`
