+++ components/script/dom/element/element.rs
@@ -99,7 +99,6 @@ use crate::dom::bindings::domname::{
 use crate::dom::bindings::error::{Error, ErrorResult, Fallible};
 use crate::dom::bindings::inheritance::{Castable, ElementTypeId, HTMLElementTypeId, NodeTypeId};
 use crate::dom::bindings::num::Finite;
-use crate::dom::bindings::refcounted::{Trusted, TrustedPromise};
 use crate::dom::bindings::reflector::DomObject;
 use crate::dom::bindings::root::{Dom, DomRoot, LayoutDom, MutNullableDom, ToLayout};
 use crate::dom::bindings::str::DOMString;
@@ -116,8 +115,6 @@ use crate::dom::domrect::DOMRect;
 use crate::dom::domrectlist::DOMRectList;
 use crate::dom::domtokenlist::DOMTokenList;
 use crate::dom::elementinternals::ElementInternals;
-use crate::dom::event::{EventBubbles, EventCancelable, EventComposed};
-use crate::dom::eventtarget::EventTarget;
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::html::htmlanchorelement::HTMLAnchorElement;
 use crate::dom::html::htmlbodyelement::{HTMLBodyElement, HTMLBodyElementLayoutHelpers};
@@ -183,7 +180,6 @@ use crate::dom::virtualmethods::{VirtualMethods, vtable_for};
 use crate::script_runtime::CanGc;
 use crate::script_thread::ScriptThread;
 use crate::stylesheet_loader::StylesheetOwner;
-use crate::task::TaskOnce;
 
 // TODO: Update focus state when the top-level browsing context gains or loses system focus,
 // and when the element enters or leaves a browsing context container.
@@ -5488,77 +5484,6 @@ impl TagName {
     /// next time that `or_init()` is called.
     fn clear(&self) {
         *self.ptr.borrow_mut() = None;
-    }
-}
-
-pub(crate) struct ElementPerformFullscreenEnter {
-    element: Trusted<Element>,
-    document: Trusted<Document>,
-    promise: TrustedPromise,
-    error: bool,
-}
-
-impl ElementPerformFullscreenEnter {
-    pub(crate) fn new(
-        element: Trusted<Element>,
-        document: Trusted<Document>,
-        promise: TrustedPromise,
-        error: bool,
-    ) -> Box<ElementPerformFullscreenEnter> {
-        Box::new(ElementPerformFullscreenEnter {
-            element,
-            document,
-            promise,
-            error,
-        })
-    }
-}
-
-impl TaskOnce for ElementPerformFullscreenEnter {
-    /// Step 9-14 of <https://fullscreen.spec.whatwg.org/#dom-element-requestfullscreen>
-    fn run_once(self, cx: &mut js::context::JSContext) {
-        let element = self.element.root();
-        let promise = self.promise.root();
-        let document = element.owner_document();
-
-        // Step 9
-        // > If any of the following conditions are false, then set error to true:
-        // > - This’s node document is pendingDoc.
-        // > - The fullscreen element ready check for this returns true.
-        // Step 10
-        // > If error is true:
-        // > - Append (fullscreenerror, this) to pendingDoc’s list of pending fullscreen events.
-        // > - Reject promise with a TypeError exception and terminate these steps.
-        if self.document.root() != document ||
-            !element.fullscreen_element_ready_check() ||
-            self.error
-        {
-            // TODO(#31866): we should queue this and fire them in update the rendering.
-            document
-                .upcast::<EventTarget>()
-                .fire_event(atom!("fullscreenerror"), CanGc::from_cx(cx));
-            promise.reject_error(
-                Error::Type(c"fullscreen is not connected".to_owned()),
-                CanGc::from_cx(cx),
-            );
-            return;
-        }
-
-        // TODO(#42067): Implement step 11-13
-        // The following operations is based on the old version of the specs.
-        element.set_fullscreen_state(true);
-        document.set_fullscreen_element(Some(&element));
-        document.upcast::<EventTarget>().fire_event_with_params(
-            atom!("fullscreenchange"),
-            EventBubbles::Bubbles,
-            EventCancelable::NotCancelable,
-            EventComposed::Composed,
-            CanGc::from_cx(cx),
-        );
-
-        // Step 14.
-        // > Resolve promise with undefined.
-        promise.resolve_native(&(), CanGc::from_cx(cx));
     }
 }
