+++ components/script/dom/fullscreen/element.rs
@@ -2,8 +2,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
+use crate::dom::bindings::error::Error;
 use crate::dom::bindings::inheritance::Castable;
 use crate::dom::bindings::refcounted::{Trusted, TrustedPromise};
+use crate::dom::document::document::Document;
 use crate::dom::element::element::Element;
 use crate::dom::event::event::{EventBubbles, EventCancelable, EventComposed};
 use crate::dom::event::eventtarget::EventTarget;
@@ -18,6 +20,77 @@ impl Element {
             return false;
         }
         self.owner_document().get_allow_fullscreen()
+    }
+}
+
+pub(crate) struct ElementPerformFullscreenEnter {
+    element: Trusted<Element>,
+    document: Trusted<Document>,
+    promise: TrustedPromise,
+    error: bool,
+}
+
+impl ElementPerformFullscreenEnter {
+    pub(crate) fn new(
+        element: Trusted<Element>,
+        document: Trusted<Document>,
+        promise: TrustedPromise,
+        error: bool,
+    ) -> Box<ElementPerformFullscreenEnter> {
+        Box::new(ElementPerformFullscreenEnter {
+            element,
+            document,
+            promise,
+            error,
+        })
+    }
+}
+
+impl TaskOnce for ElementPerformFullscreenEnter {
+    /// Step 9-14 of <https://fullscreen.spec.whatwg.org/#dom-element-requestfullscreen>
+    fn run_once(self, cx: &mut js::context::JSContext) {
+        let element = self.element.root();
+        let promise = self.promise.root();
+        let document = element.owner_document();
+
+        // Step 9
+        // > If any of the following conditions are false, then set error to true:
+        // > - This’s node document is pendingDoc.
+        // > - The fullscreen element ready check for this returns true.
+        // Step 10
+        // > If error is true:
+        // > - Append (fullscreenerror, this) to pendingDoc’s list of pending fullscreen events.
+        // > - Reject promise with a TypeError exception and terminate these steps.
+        if self.document.root() != document ||
+            !element.fullscreen_element_ready_check() ||
+            self.error
+        {
+            // TODO(#31866): we should queue this and fire them in update the rendering.
+            document
+                .upcast::<EventTarget>()
+                .fire_event(atom!("fullscreenerror"), CanGc::from_cx(cx));
+            promise.reject_error(
+                Error::Type(c"fullscreen is not connected".to_owned()),
+                CanGc::from_cx(cx),
+            );
+            return;
+        }
+
+        // TODO(#42067): Implement step 11-13
+        // The following operations is based on the old version of the specs.
+        element.set_fullscreen_state(true);
+        document.set_fullscreen_element(Some(&element));
+        document.upcast::<EventTarget>().fire_event_with_params(
+            atom!("fullscreenchange"),
+            EventBubbles::Bubbles,
+            EventCancelable::NotCancelable,
+            EventComposed::Composed,
+            CanGc::from_cx(cx),
+        );
+
+        // Step 14.
+        // > Resolve promise with undefined.
+        promise.resolve_native(&(), CanGc::from_cx(cx));
     }
 }
